* Consolidate authManager and c11nboot * Apply latest auth changes from other sdks
